Refactor adapter registration for resources
The same code to register adapters for resources is now called for default
resources as well as custom resources. If custom resources are registered
from a plugin, the plugin must make sure to add the appropriate
`mako.directories` settings before calling `config.register_resource` if it
provides templates for adapters.

Closes #75